Upversion to 0.15.0-alpha.1
Upgrading cert-manager to v0.15.0-alpha.1 in order to fix functional
issues noticed with previous version.

This commit also includes affinity fix in manifest file to deploy
pods on controller.

This commit also removes CRD definitions deployed as second chart.
It was noted that this was the cause of functionality issues, so making
changes to build system to copy CRDs to templates/ folder and using
installCRDs option in manifest file to fix the issue.

Tested by applying issuer & generating certificate.

Story: 2007360
Task: 39222

Change-Id: I2ecb425c1ef2d87069df786f43a9d1ecd7fe5435
Signed-off-by: Sabeel Ansari <Sabeel.Ansari@windriver.com>