Enable helm plugin delivery with the application
This creates a new package spec called python-k8sapp-cert-manager that
will build two packages: python-k8sapp-cert-manager and
python-k8sapp-cert-manager-wheels. These packages will hold all the
stevedore plugins needed to support the application.

These packages are included in the build dependencies for the
stx-cert-manager-helm package build where the wheels file will be
included in the application tarball.

The helm plugins have been relocated to this repo and provided in a
k8sapp_cert_manager python module. This module will be extracted from
the wheels and installed on the platform via the sysinv application
framework. The module will be made available when the application is
enabled.

Now that the plugins and the application are residing in the same
repository, GITREVCOUNT will be used to drive the package version of
both to keep them in sync.

Tox test support has been provided for the plugin module.

Change-Id: I13cb0f7d429507d1b1d6bdc16cdd3750f370c8e0
Depends-On: https://review.opendev.org/#/c/734010/
Story: 2006537
Task: 39992
Signed-off-by: Robert Church <robert.church@windriver.com>