add extraArgs enable-certificate-owner-ref
When removing an application (e.g., vault) that had provisioned a
certificate with cert-manager, "the issuer and certificate resources are
deleted, but the secret is left behind because it is not owned by the
certificate resource... This orphaned secret does not get updated on a
subsequent reapply and is then invalid when vault attempts to use it."

The workaround was to remove the secret manually before reapply.

Enable the option by default for StarlingX. Secrets will be
automatically removed when the certificate resource is deleted.

Closes-Bug: 1888900
Change-Id: I2b057a71da8dd761a891fc879ad9860c9822cba0
Signed-off-by: Michel Thebeau <Michel.Thebeau@windriver.com>