Fix pylint and py27 failures in zuul
pylint running in python2.7 will complain that interperter
can not be found.

Rather than running the pylint job run in python2.7, bump
the job to python3.6, and set the node to run on ubuntu-bionic.

py2.7 jobs was failing randomly because pbr was being setup in the
sdist step. It is unclear if the reason is that something is
source packaged during tox which conflicts with another zuul job
of a different interperter, so simply disable to the source
dist generation.

Most openstack projects are configured like this as well.

Copyright was updated so that the gate will pick up this change.
Closes-Bug: 1932186
Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: Ia677057690d3ffe87c23471d958dbff670db9b53
(cherry picked from commit 8253ea8ed033078a2f4510f963b0ca2494b1ab65)