Remove armada specific resources and build
This deletes armada left over resources and updates the centos and
debian packaging scripts to remove the armada package build.

This is also deleting a cert-manager-helm child package which is not
needed anymore. Prior to cert-manager 1.7.1, it was used to extract
the upstream helm-chart, apply patches and helm-package again. The
updated helm-chart would be then used by stx-cert-manager-helm to
generate the armada tarball.

That two-steps approach was dropped in the upversion to 1.7.1 but the
folder was kept. This change is cleaning that up as well.

Test Cases:

PASS: Centos build and iso install.
PASS: Debian build and iso install.
PASS: Simple functional test creating cert-manager resources.
PASS: Verified that toleration patch still works.

Story: 2009138
Task: 45678
Story: 2009837
Task: 45256

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I9d36c8377e7893e09e6adae16e8defbad0dd824f