switch to use new K8s control-plane label
The node-role.kubernetes.io/master= label is being deprecated and is
being replaced by node-role.kubernetes.io/control-plane=

Similarly the node-role.kubernetes.io/master:NoSchedule taint is being
replaced by node-role.kubernetes.io/control-plane:NoSchedule

We need to update the code appropriately.  We can switch to the new
label since it's already available, but where we have a toleration for
the existing taint we need to keep the existing toleration and add
another one for what will be the new taint.

Test Plan:
PASS: Validate that application is applied successfully.
PASS: Validate that new nodeSelector is used and pods are running
      on control plane nodes.
PASS: Validate that new toleration is applied.

Story: 2010301
Task: 46574

Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: Ib09bcdc37cd6c2fb3964f5c87eea7caa7a0d52d3