Fix zuul failures during setup
The ubuntu-jammy nodeset gets selected by default
and is causing problems during setup.

Collecting cffi>=1.1
Failed to build cffi

ubuntu-focal seem to work fine.
Will specify the nodeset to be focal to resolve this.

Need to update a file that is monitored by zuul
in order to trigger the failing zuul jobs.

Partial-Bug: 1994843

Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Iefa7fcbf3c293b10fca204e7c79a4e3b3d5d5ae3