Update debian packages for pkg-versioning
The Debian packaging meta_data file has been changed to reflect all the
latest git commits under the directory, pointed as usable, and to
improve pkg-versioning addressing the first commit as start point to
build packages.

This ensures that any new code submissions under those
directories will increment the versions.

The commit SHA df8fc0aa19 was chosen to be the SRC_GITREVCOUNT of the
python3-k8sapp-cert-manager's metadata because is the commit that
creates the package file structure in the repo.
(python3-k8sapp-cert-manager version was modified from .stx.60 to
.stx.61)

The commit SHA afd731d65d was chosen to be the SRC_GITREVCOUNT of the
stx-cert-manager-helm's metadata because is the commit that creates
the package file structure in the repo.

The commit SHA 78cacfb0d7 was chosen to be listed in FILES_GITREVCOUNT
because is the commit that adds the "helm-charts" directory of the
repository referenced in the "src_files" for stx-cert-manager-helm's
metadata.
(stx-cert-manager-helm version was modified from .stx.60 to .stx.61)

stx_patch is added to revision section to maintain the up versioning
of both packages.

Test Plan:
PASS: Verify package versions are updated as expected.
PASS: build-pkgs -c -p python3-k8sapp-cert-manager
PASS: build-pkgs -c -p stx-cert-manager-helm

Story: 2010550
Task: 47495

Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>
Change-Id: I1b15086b117889471dfa859d8d9a597ec060e4ea