cert-manager: Up-version to version: 1.11.1
This commit upversions the cert-manager helm charts to v1.11.1, which is
the latest version at the time. This commit also align the patches to
the new helm charts.
- the issue addressed by patch 0001 has been merged upstream, so the
patch was reworked to maintain only the addition of the Chart.yaml file.

Test Plan:

PASS: Create a certificate managed by cert-manager and verify if it
      automatically renewal in the right time.
PASS: Create certificate with 30 or less days to expire and verify
      that the "expiring soon" alarm is created.
PASS: Create certificate and wait until it is expired and verify if
      "expired" alarm is created.
PASS: Apply applications(Portieris, OIDC and Vault) and verify if it
      applied succesfully.
PASS: Delete a secret of a certificate and verify if cert-mon was able
      to reinstall it.
PASS: Run full build, system install, bootstrap and unlock(SX and DC).
PASS: Delete the secret of subcloud certificate and verify if cert-mon
      was able to reinstall and replicate it to the subcloud system.
PASS: Create a local intermediate CA, self-signed and ACME issuers and
      verify if the secrets and the protocols are running succesfully.

Depends-on: https://review.opendev.org/c/starlingx/ansible-playbooks/+/882730

Story: 2010704
Task: 47986

Change-Id: Id7f29263582faa030156d0589f12880083d31d7f
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>