Enable automatic app update for cert-manager
This commit enables cert-manager-armada-app to be automatically
updated. The auto update will be triggered only when the app is in
applied state and a tarball with an increased version is present
on the system. Only one tarball per app should be present on the
directory, so the automatic update can be performed.

Test Plan:
PASS: Run full build, system install, bootstrap and unlock SX system,
      upload a tarball file with an upper version than the actual to
      the helm applications directory and observe cert-manager
      automatically update.
PASS: Build stx-cert-manager-helm package with "auto_update: true" and
      manually apply the app in the system, then upload a tarball file
      with an upper version than the actual to the helm applications
      directory and observe cert-manager automatically update.
PASS: Build stx-cert-manager-helm package with "auto_update: true",
      then upload the tarball file to the helm applications
      directory and observe cert-manager automatically update.

Story: 2010628
Task: 48158

Signed-off-by: Manoel Benedito Neto <manoel.beneditoneto@windriver.com>
Change-Id: I769148091fa6fd456cabe4ec733c1922616755b4