Issue test certificate via lifecyle
Issue a test certificate using the application lifecycle in order to
assert cert-manager readiness. This allows dependent applications to
promptly rely on cert-manager once it is applied.

This is a re-implementation of the original code, which was removed with
the 21-k8s-app-upgrade.sh script, in order to allow apps to be updated
directly via the Application Framework using the new dependency feature:
https://review.opendev.org/c/starlingx/update/+/953526

Tox sysinv import was fixed for the application plugins as part of this
change.

Test plan:
PASS: build-pkgs && build-image
PASS: upload/apply/update/remove/delete cert-manager
PASS: Platform fresh install
PASS: Apply ptp-notification-app
      Platform upgrade from stx 10

Depends-on: https://review.opendev.org/c/starlingx/config/+/956724

Closes-bug: 2119681

Change-Id: Ic37e97ee47c08eabcb1afde5c2a820bf68196326
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>