Increase cert-manager qps and burst limit
This change added --kube-api-qps=50 and --kube-api-burst=100 to
containers' args of cert-manager's deployment. This override the
default qps (reqests per second limit) and burst (request burst limit)
for cert-manager, which are too low (5 and 10 respectively) and causing
cert-manager to retry and requeue requests when cert-manager is
handling hundreds of certificate requests in burst in cases such as
large scale DC system deployment.

Test Plan:
PASS: In a DX system, update cert-manager application with the change,
      verify cm-cert-manager deployment has the two args added:

      containers:
      - args:
        - --kube-api-qps=50
        - --kube-api-burst=100

PASS: Use a script to simulate subclouds creation in a batch of 250,
      loop through the batch creation for multiple times (~27 times).
      Verify cert-manager doesn't generate the following logs:

      "Waited for 1.293433266s due to client-side throttling, not
       priority and fairness."
      "re-queuing item due to optimistic locking on resource"

PASS: Fresh install a DX system, verify deployment is successful, in
      particular cert-manager behaves as expected with the change.

Closes-Bug: 2127880

Change-Id: Ie5687b729956e9ab1a56c91699b09c11f2011093
Signed-off-by: Andy Ning <andy.ning@windriver.com>